requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
Flask==1.1.4
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pygame==2.0.3
python-dotenv==0.19.1
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
waitress==2.0.0
Werkzeug==1.0.1
